requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
coloredlogs==15.0.1
frozenlist==1.8.0
humanfriendly==10.0
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.7.0
neotermcolor==2.0.11
propcache==0.4.1
psutil==7.2.1
py-cord==2.7.0
pyfiglet==1.0.4
Pygments==2.19.2
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.6.2
yarl==1.22.0
